Moved each of the classes into its own file
